+++ src/main.lua
@@ -46,7 +46,11 @@ local function on_ready()
     if config.enabled == false then return end
     mod = modutil.mod.Mod.Register(_PLUGIN.guid)
     mod.config = config
-    mod.IsZag = rom.mods["NikkelM-Zagreus_Journey"] and rom.mods["NikkelM-Zagreus_Journey"].config and rom.mods["NikkelM-Zagreus_Journey"].config.enabled and config.zags_journey_integration
+    mod.IsZag = rom.mods["NikkelM-Zagreus_Journey"] and
+                rom.mods["NikkelM-Zagreus_Journey"].config and
+                rom.mods["NikkelM-Zagreus_Journey"].config.enabled and
+                rom.mods["NikkelM-Zagreus_Journey"].IsValidInstallation and
+                config.zags_journey_integration
 
     mod.PostSetupRunDataFuncs = {}
     mod.PreSetupRunDataFuncs = {}